shop-srv/user_srv/requirements.txt
peewee
pymysql
passlib
grpcio
grpcio-tools
protobuf
loguru
python-consul2
redis
dnspython
nacos-sdk-python
rocketmq-client-python
jaeger-client
ratelimit
pybreaker